Refactor s2i Tests for ci-operator
* Isolate integration tests that don't require Docker
* Add OWNERS file
* Add make commands to run integration tests
