+++ test/integration/docker/doc.go
@@ -1,0 +1,2 @@
+// Package docker contains docker-related integration tests for S2I.
+package docker
