+++ test/integration/dockerfile/doc.go
@@ -1,0 +1,1 @@
+package dockerfile
